cleaned up my data a bit more
alt_run_e is just alt_run again to make things easier to keep track of. I also added two new columns which show the cycles that are switch trials (when they switch from one task dimension to another) and which cycles are control trials. Will be working on this just to see if we could get some more useful data for our graphs but this is just extra.